// src/shared/models/sahred-product-translation.model.ts
import { z } from 'zod'

export const ProductTranslationSchema = z.object({
  id: z.number(),
  productId: z.number(),
  name: z.string().max(500),
  description: z.string(),
  languageId: z.string(),

  createdById: z.number().nullable(),
  updatedById: z.number().nullable(),
  deletedById: z.number().nullable(),
  deletedAt: z.date().nullable(),
  createdAt: z.date(),
  updatedAt: z.date(),
})

// TYPE EXPORT

export type ProductTranslationType = z.infer<typeof ProductTranslationSchema>